Extract new task to separate component
